Specify a default 20 second timeout for requests
The packaged FPM check uses an unspecified
[timeout](http://docs.python-requests.org/en/master/user/quickstart/#timeouts)
for requests it makes, which can potentially lead to failing pools
taking forever to return a response (or as long as intermediary
services/proxies/etc allow).

The effect is that should a pool fail to return a response, the
particular check can hang, blocking execution of _all_ further checks
and subsequent metric submission.
https://github.com/DataDog/dd-agent/issues/2854 is a useful reference
point for a similar issue.

This can cause the agent
[watchdog](https://github.com/DataDog/dd-agent/wiki/FAQ#why-is-my-agent-process-dying)
thread to kick in and terminate the agent before any data is returned.

Ultimately, this means one failing pool can block the return of any data
at all.

A 20 second timeout seems a reasonable compromise for an ill performing
pool to prevent this scenario.  Should the operator require, they can
specify a custom timeout.