User-defined "url" tag replaces default "url" tag.
Using service discovery with docker the IP addresses of the url may change often resulting in a check going down with one IP and coming back up with another which causes the alerting system to not be able to auto-resolve the alert. With this change if the user defines their own "url" tag the change of IP won't matter and the check will be able to auto-resolve properly.