[mongo] Sanitize server URI correctly
Changes the approach to sanitize the server URI so that it works well
whether the username and/or password are url-encoded or not in the
raw server string.

Added unit tests to take into account all known cases.

Also, removed the `requires=mongo` attribute on the unit tests
of the integration.