Support fetching Kafka consumer offsets stored in kafka
This is a functioning check that fetches consumer offset stored in
Kafka. Note that this is different than the old consumer offsets that
are stored in Zookeeper. This is a work in progress as some of the
error-handling is incomplete. Additionally, it still needs to get the
kafka 0.10.2 support added.