[ssh_check] Fix thread leak
Since the upgrade to paramiko 2.x, it appears that a failure to close
connections leaks threads. To fix this, make sure we always close
connections.

We always close the client, even when the connection is unsuccessful
(paramiko handles this case gracefully).

Add an assertion on the number of threads after the check has run.

Also, remove the mention of a limit on the size of `ecdsa` keys in
the example yaml file, since the limit applies only to paramiko 1.x.