[zk] Remove usage of `set` (#473)
`set` is deprecated in the Agent. The usage of `set` in the `zk` check
doesn't make sense anyway since for a given set of tags the metric is
always either submitted with the same value (`self.hostname`), or not
submitted at all. So it either reports `1` or doesn't report
anything. Using `gauge` yields the same result, so let's use `gauge`
instead.

Updates changelog and manifest too