[openstack] Adds human friendly "project_name" tag in all cases.
If the project is specified in the check config using the tenant id (rather than using the domain id and the project name), the check will now attempt to fetch the project name from the keystone API, so that it may be added to project and server metrics as the "project_name" tag. Previously, if the project was specified using the tenant id, the only tag that could be used to differentiate projects was the tenant_id tag, which isn't easy for humans to understand.