[kafka_consumer] discovery of groups, topics and partitions (#631)
* Make args optional in kafka_consumer.py

The consumer_groups, topics, and partitions arguments are each now
optional. Any omitted values will be fetched from Zookeeper. If a value
is omitted, the parent value must still be it's expected type (typically
a dict).

Modern Kafka consumers store their offset in Kafka rather than
Zookeeper, but this PR does not add support for this. It's tricky
because this check is much simpler if it assumes a single place can be
queried for all consumer offsets, but currently there's no easy way to
do that. Once KIP-88 is implemented, it will be much easier to add this
functionality, although it would only work for Kafka brokers >=
0.10.2.0. In the meantime, you can instrument your individual kafka
consumers to submit their offsets to Datadog.

* Add monitor_unlisted_consumer_groups flag which defaults to False

* [kafka_consumer] send event if consumer lag negative.

* [kafka_consumer][test] add a test to auto_discover ZK groups.

* [kafka_consumer] fix event submission.

* [kafka_consumer] format missing consumer_group key.

* [kafka_consumer] fix all path definitions + test.

* [kafka_consumer] modifying aggregation key.

* [kafka_consumer] updating changelog and manifest.