Send CRITICAL service check for aliveness if the rabbitmq server goes down
Since service_checks on the backend keep the latest value they received, we would still show a OK service_check for vhost
aliveness even if the server is down. We now cache the latest list of vhosts we pulled from rabbitmq to send CRITICAL checks
if the server don't respond anymore.