Clarify DogStatsD vs. StatsD, add more info on DogStatsD
Introduce and explain what DogStatsD is before referencing it in the text. The fact that there is a link out to the gunicorn docs first that talks about StatsD and no explanation about DogStatsD makes it not clear whether StatsD upstream needs to be installed (it doesn't need to be).