Updating excluded_disks example
The disk.yaml file provides no Windows example. However, the name of disks in Windows is not particularly obvious as it needs to include the backslash of the root directory. Because of this, users are often having to resort to complex regexes (set in excluded_disk_re) instead of the simple excluded_disks section