Adding wiredTiger page read/write data to metrics
wiredTiger page read and write data is critical data, useful for performance tuning. I think it should be included in the monitored metric list.

On a side note maybe we should also add bytes/read and bytes/written to the monitored metrics?

I added the #noqa comment to indicate that the collection of this metric doesn't have an integration/unit test. Was I correct in doing so?
Fix of a syntax error

Fixing a syntax error in mongo/check.py