[consul] Add support for consul 1.0.0
Consul 1.0.0 is the first version that Hashicorp considers "stable" in
terms of API calls. As such, previous versions of consul may return
different payloads, and the API calls that the Datadog agent currently
uses to determine cluster leadership no longer work on 1.0.0 and later.

The current method of determining cluster leadership is to retrieve the
local agent's IP:Port combination and compare it to the IP:Port of the
leader. While consul 0.7.0 provides a more robust way of determining
leadership (literally a true/false key), this method is not available
in consul 0.6.4, so using it would break compatibility. Therefore this
commit makes the smallest possible change to avoid breaking that
compatibility, by instead using a different method of fetching the
local agent's IP:Port combination.

Tests have been updated (and somewhat refactored) and now supports
versions 0.6.4, 0.7.2, and 1.0.0.