Adds the option to send the raw deltas of IIS WMI performance counters.
Previously, the deltas of these performance counters were normalized to per-second rates, then submitted as gauges. The agent does not run checks are perfectly regular intervals, so it isn't possible to exactly derive the raw counts from the rate metrics. Furthermore, forcing a gauge to behave like a count requires advanced knowledge of the query system.

The agent did not have the monotonic_count method when this integration was changed to use the rate method:
https://github.com/DataDog/dd-agent/commit/d00a5e30bfd4542e37c95305a4aaf409e316c4d6
https://github.com/DataDog/dd-agent/pull/1012/files

This commit adds the count metrics with new metric names to allow the metadata to be updated without breaking existing agent deployments, and these new metrics are behind an option that is off by default. This commit also adds an option to control whether the rate metrics are sent. Some users may wish to keep the old metrics for backwards compatibility or to complement the raw count metrics.