Remove deprecated _read_config()
A while back this was deprecated because it will not be used in the new
version of the agent. At the time, this was handled by copy/paste the
deprecated method into this check.

However, this method should actually be completely removed, as it adds
no value to the check and is easily refactored away.