Let the client intelligently pick a random node
Previously, a node was chosen at random. However, kafka-python offers a
built-in method to intelligently pick a node, preferring nodes that are
already connected (lower latency, less error-checking), with the fewest
number of outstanding requests (lower latency), and only if no brokers
are already connected will it fallback to a random node.

Note: This removes a custom exception, however the exception is not
actually needed because _make_blocking_req() calls _ensure_ready_node()
which already doublechecks the node is ready.