PR changes - tag cardinality, docs, code refactor
Reduce the cardinality of the tags for the pod.status_phase metric to
increase scalability.

Rework the pod.status_phase metadata docstring to better represent what
the metric is and how to use it.

Refactored the code to make use of existing constructs.

Adds "phase" tag to the service check portion of the pod_status_phase
prometheus metric processing. Update tests to match.