Cleanup group coordinator lookup
You must have a known group coordinator to lookup the offsets for a
consumer group that stores its offsets in Kafka. I do not understand
why this code tries to handle unknown group coordinators by
firing requests to all the brokers as that seems like a waste of network
calls.

This only refactors the existing code to be cleaner. There is room for
additional error handling and logging. Plus potentially caching the
group coordinator value on the client's cluster metadata rather than
re-fetching it every time... instead just assume it's still the old one,
and if not catch the error and trigger a group coordinator lookup.

But that is left for a later PR, this one merely cleans up the existing
logic to make it easier to refactor later.