Fix bug parsing squid metrics
Squid can return a trailing newline at the end of its metrics. Prior to
this change the trailing newline would be parsed as a metric line, and
an error would be raised due to the missing "=" character. This strips
the trailing newline from the metrics before they're parsed line by
line.