Add sister gauge metrics to kubernetes_state pod service checks (#1578)
* Add kube_pod_status_phase as gauge

It was previously only a status check which is not helpful if you want to
visualize the status of your pods over time.

* PR changes - tag cardinality, docs, code refactor

Reduce the cardinality of the tags for the pod.status_phase metric to
increase scalability.

Rework the pod.status_phase metadata docstring to better represent what
the metric is and how to use it.

Refactored the code to make use of existing constructs.

Adds "phase" tag to the service check portion of the pod_status_phase
prometheus metric processing. Update tests to match.

* Make pod.status_phase metric report value from prometheus

* Rework pod.status_phase check

Now submits an agent local count tagged with the five pod phases and
with the different namespaces with pods. This should allow tracking
aggregates of these metrics cluster-wide while retaining performance.

Updated tests and description to match.

* Update pod.status_phase metric submission to aggregate client side counters