Adds in the ability to specify 'use_cached' in the pgbouncer config.
During our use of the datadog pgbouncer check, we found an issue where the agent would fail
to reconnect to a recovered pgbouncer process.  We simulated this by starting our
`pgbouncer` and `datadog-agent` processes, and then subsequently stopping `pgbouncer`,
waiting a few minutes, and restarting `pgbouncer`.

The agent would correctly alert us that `pgbouncer` was down, but upon recovery, the agent
would continue to be in an error state.  We would have to restart the agent to get it reporting "healthy" again.
We have reason to believe this is due to the caching of the underlying Postgres connection.

We simulated this again with the change in this PR and the agent behaved as expected.

Please let us know if there's additional testing coverage or simulations we can provide.

Cheers!