Include ContainerCreating in k8s pod waiting status reasons
I've seen different issues where the container can get stuck in the
ContainerCreating status without moving to any of the other listed waiting
reasons. For example, a bug in [amazon-vpc-cni-k8s][1] can cause a pod to get
stuck in that status if the CNI plugin is unable to assign an IP for the pod.
Whitelisting this status allows monitoring for these cases.

With this reason added to the list, the list now contains all the [possible
reasons][2] and is therefore unnecessary barring future addition of reasons. If
it were up to me, I'd simply drop the whitelist check, but there seems to be
[some opposition][3] to that.

[1]: https://github.com/aws/amazon-vpc-cni-k8s
[2]: https://github.com/kubernetes/kube-state-metrics/blob/bcb230560fe1a32810577762f0b91ad0a5f50d1a/Documentation/pod-metrics.md
[3]: https://github.com/DataDog/integrations-core/pull/1763#issuecomment-407709937