Bump kafka-python to 1.4.4
This bumps to 1.4.4 which added the new `KafkaAdminClient` which we need
in order to make `monitor_unlisted_consumer_groups` work for consumer
offsets stored in Kafka.

It also comes in handy since it will let us remove a bunch of code from
the existing `KafkaCheck` and just rely on upstream.