Add support for fetching unknown consumer offsets from Kafka
This adds support for dynamically discovering/fetching metrics for
unlisted consumer groups that store their offsets in Kafka.

Previously, `monitor_unlisted_consumer_groups` only worked for
consumers offsets stored in Zookeeper.

There are a number of edge cases to adding this feature, see the full
decision tree in `conf.yaml.example` to understand how the check decides
which offsets to fetch from which source.

In order to accomplish this, I had to switch from
`kakfa-python.KafkaClient` to the new [`kafka-python.KafkaAdminClient`](https://kafka-python.readthedocs.io/en/master/apidoc/KafkaAdminClient.html).
In addition to allowing new functionality, this also resulted in
removing a lot of code that was now superfluous.