add testing with tox and fix some flake8 errors
NOTE: Due to the `usedevelop=true` setting with tox (in order to
accommodate reading `requirements.in`), metadata and targets may be
written to current development environment. This is not an issue with
running tox tests on the cloud, but it may be an issue with developer
environments. We use .gitginore files to try to make sure that
developers do not accidentally check in unintended metadata and targets.