Add datadog-checks-downloader (#3026)
* Add integrations-core-downloader.

* add prodsec as code owner

* add a useful comment

* fix comment indentation

* remove TODOs we are not addressing for now

* restructure directories per @gzussa feedback

* add testing with tox and fix some flake8 errors

NOTE: Due to the `usedevelop=true` setting with tox (in order to
accommodate reading `requirements.in`), metadata and targets may be
written to current development environment. This is not an issue with
running tox tests on the cloud, but it may be an issue with developer
environments. We use .gitginore files to try to make sure that
developers do not accidentally check in unintended metadata and targets.

* pin requirements

* update README

* `ddev dep freeze`

also some minor improvements in documentation

* run pytest with coverage

* add ProdSec on the signing tool

also add .coveragerc

* increase coverage bar

* make a method private

also reduce coverage bar because local < remote for some reason TBD

* logging output now works thanks to @ofek

* address @zippolyte feedback

* address @zippolyte feedback

* remove dependency on `click`

also fix logic in logging level

* A few improvements:

* Update agent requirements
* Simplify tox environments
* Check that given check version matches PEP 440

* bring back python 3.6 because travis

* update setuptools and fix doc

* update agent requirements

* remove setuptools as a dependency

* assume setuptools is installed by default

* remove unnecessary coveragerc

* remove install_requires from setup.py

also increase coverage by a bit

* Apply suggestions from code review

Co-Authored-By: trishankatdatadog <33133073+trishankatdatadog@users.noreply.github.com>

* incorporate @ofek feedback

* address @ofek feedback

* address @ofek feedback