Use `{proc_location}/net/stat/nf_conntrack' for `conntrack.entries'.
Reading the number of lines from `{proc_location}/net/nf_conntrack' for
the metric `system.net.conntrack.count' is really slow when the number of
`conntrack'entries are large. It issues a lot of `read()' calls to the
`nf_conntrack' file on `procfs' and the collector agent "self-destructs"
(probably killed by `watchdog') while waiting to fetching all the entries
from the file.

An alternative is to use the `{proc_location}/net/stat/nf_conntrack' file,
which has a summarized list of entries in the first column for each processor,
which makes it a fewer lines to read.