Adjust latency tracking in redisdb integration (#3689)
In my Redis monitoring, I see extremely frequent latency spikes to ~100ms (leading me to believe there's a timeout of _something_ being hit). After fairly significant investigation, I can't find any indication from Redis itself that these are genuine. That led me to believe that the latency reported was in the actual monitoring agent.

As such, this proposal reduces the amount of activity that happens inside the latency timer to just the `conn.info()` call. It at least gives me more confidence that the measured Redis latency is accurate, and not including the submission time of other metrics.