WIP: Automated E2E testing framework
This adds a new test subcommand to run a full E2E testing with an agent.
It adds new pytest marker which will use the check output to populate
the aggregator.

This is tested against lighttpd. To check how it works, you can call
`ddev env test lighttpd py27-latest`.