Deployment environment with Terraform
This adds the ability to deploy an environment for e2e with Terraform.
It adds the environment runner, and the ability to create a SOCKS tunnel
to the instance using SSH to run the check against the environment.

I implemented the environment for the OpenStack check, and SOCKS first
instead of regular tunnel because the OpenStack check talks to several
endpoints.

To test, you need to download a GCP credentials files and set the
`TF_VAR_account_json` environment variable.