Add e2e tests
This adds the first E2E test for a JMX check. It works around the small
differences in output, and change the behavior of `ddev env check` to
call `agent check` by default for JMX. It also deals with the
particuliar JAR loading for JBoss.

The `jmx list` capability is retained through the `--jmx-list` option of
the check command.