Remove rarely used zookeeper-specific min collection interval
This was a feature I'd pushed for several years ago to support fetching
consumer offsets from Zookeeper at a different rate than from Kafka.

However, now that Agent v6 supports setting `min_collection_interval` at
the instance level, this is no longer needed... instead, just configure
two separate instances of the check--one for kafka and one for
zookeeper.

So removing this code to simplify the legacy check and making the
setting a no-op.