Stop documenting `min_collection_interval` at `init_config` level
In Agent v6, this is now set at a per-instance level rather than in
`init_config`.

Given that the Zookeeper functionality is deprecated and fetching
consumer offsets from Kafka should be a fast / low-load operation, I
think it best to remove the documentation entirely rather than moving it
down to the `instance` level.

If someone wants it for some reason, it's easy to look up.