Cleanup pointless retries that query the wrong brokers
The retry functionality makes very little sense here. Exponential
backoffs are already built into the `kafka-python` client. And the rest
of the retrying... when I dug into it, is actually "retrying" by
pointlessly querying the wrong brokers.

So I cleaned it up. The timeouts within `kafka-python` will handle
killing the requests after a certain amount of time, and it's safe to
call `poll()` repeatedly because under the covers it's just doing a
local `selectors` call.