Simplify KafkaClient handling
Agent v6 only has one instance per instances, so we don't need to have a
separate cache... just cache the client on the instance.

This also lets us be sure we are closing the client at the end of the
check to avoid stale connections on older kafka brokers.

This cleanup will make life simpler when the check is refactored as part
of #3957.