[datadog_checks_base] Move checks from datadog-agent cmd/agent/dist folder
Some checks remained in the cmd/agent/dist folder of the datadog-agent
repository, which are installed in /opt/datadog-agent/bin/agent/dist/checks.
If the Agent is run with Python 3, a __pycache__ directory is then created
there, and is not removed when the Agent is uninstalled, which shouldn't
happen (the whole /opt/datadog-agent folder should be erased).

To fix this, these checks are moved to the equivalent datadog_checks_base
folder.