Drop iteritems & text_type
With the impending move to python 3, drop `iteritems` in favor of just
using `items`. The perf hit is minimal given the size of the data, and
just makes the code cleaner now that my eyes are used to python 3 code.

Additionally, `text_type` doesn't really make sense here... it's already
a string, no need to recast to a string.