Heavily refactor check
A large refactor, most in preparation for the new version of the check.

1. Stop pointlessly passing groups/topics around. This is much cleaner
handled at the object level no point in passing most of this around.
2. Add a flag `monitor_all_broker_highwatermarks` that allows fetching
broker offsets for topics that are not part of a consumer's
subscription.
3. Cleanup the tests slightly as one test was shoehorned onto another
test and the shared state was causing some weird breakage.
4. Streamline the overall code flow so that its more readable.
5. Change the way the context limit is enforced so that it's applied to
the sum total, rather than piecemeal. Additionally, because it's applied
to the total rather than the pieces, after discussion with @ofek I
bumped the context limit to 500. This should be a reasonable default to
prevent abuse while still handling most cases.
6. Change so that internal topics like `__consumer_offsets` are never
excluded from the broker highwater mark fetch. It's generally pointless
to fetch these as users care about the contents of the messages in that
topic, not the broker offsets of that particular topic.