Drop `'source:kafka'` from tags. (#4400)
Going forward, there will be only one cannonical source of Kafka. If
someone is needing to disambiguate, we keep the old `'source:zk'` tag.
But drop the `'source:kafka'` tag as it is not needed in the new code
path, so the simplest thing is to make the new/old code paths the same
so that the test path is identical and the codepath is also simpler.