Add option to activate namespace matching logic for openmetrics tags if namespace is present
Currently we are mixing up tags coming from objects having the same name in different namespaces when data are gathered from anything running in Kubernetes.
Using it in KSM to fix this issue.

Add the possibility to exclude some labels from being matched in the same namespace (i.e. Kubernetes Nodes, which are not NS resources)