Read udev disk labels from the blkid cache file
The default implementation to get the udev disk labels by executing
the blkid command. However, this commands requires to have access
to a couple of files in /dev and /proc in order to work correctly.
When the agent runs into a container, it does not necessarily have
/dev and /proc mounted, and mounting them recursively is not
always wanted.

This implementation uses the blkid cache file, as described in
blkid(8) and libblkid(3).