Fix TypeError where self._excluded_ifaces is None
The _submit_devicemetrics function in network.py assumed that
self._excluded_ifaces is a list and attempts to iterate over it cecking
to see if the iface var is in the list.

In class Network, the check function does the following:
self._excluded_ifaces = instance.get('excluded_interfaces', [])

The current implementation will set self._excluded_ifaces to None in the
following scenario: instances = {'excluded_interfaces': None}

Since excluded_interfaces is found within the dict, we end up with None
as the value for self._excluded_ifaces instead of an empty list which
causes the loop in _submit_devicemetrics to fail on a TypeError since it
is not list.

This change ensures that self._excluded_ifaces is always a list.