haproxy: Add option to skip datapoints during restarts
When restarting or reloading haproxy, there is a period of time where both the
previous and the current instance of haproxy are running concurrently. During
this period statistics data isn't reliable as it describes only one of the
instances.
This adds a configuration option to skip sending datapoints during a period at
the initial start of haproxy, to avoid this.