Add an option to skip reporting during restarts (#5571)
* haproxy/tests: Enable available socket tests on macOS

macOS has socket support, and most tests can run just fine there.
One exception is for unix sockets, because docker (where the haproxy fixtures
run) uses linux VMs. AF_UNIX sockets from linux and mac are not compatible
(https://github.com/docker/for-mac/issues/483)

* haproxy: Add option to skip datapoints during restarts

When restarting or reloading haproxy, there is a period of time where both the
previous and the current instance of haproxy are running concurrently. During
this period statistics data isn't reliable as it describes only one of the
instances.
This adds a configuration option to skip sending datapoints during a period at
the initial start of haproxy, to avoid this.