Prevent out of bounds on systems with an odd number of counter strings
In my system (Windows Server 2016) the length of val is  13033. This causes winpdh.py to error out (caught on line 42) due to indexing val 1 beyond it's size.

Not sure of the etiquette around submitting pull requests - but this request checks that we don't index beyond the capacity of the string-array.