Allow proxysql checks to specify stats database name
The previosuly-hardcoded stats database is correct if you're connected as an admin user, but doesn't exist if you're connected as a stats-only user.

From a security point of view, it's probably better for datadog to connect as a stats user, since those aren't granted access to sensitive data. The existing checks should all work as a stats user, but for those users the 'stats' database is called 'main' instead.

This change drops the hardcoded database name from queries and specifies a default database at connection time instead, with the default set to 'stats'. This should be backwards-compatible with the old behavior.