Conflicting Docs
It looks like on https://app.datadoghq.com/monitors#create/process that we state you use conf.d/process.yaml, and now I'm not sure if I know which one is the correct one. I assume this one is, but that means we need to update the app documentation to reflect this correctly.