[disk] Report read_time and write_time as a count of millis
The existing metrics `read_time_pct` and `write_time_pct` were not
reporting a percentage as expected (it's an always-increasing value in ms,
dividing by 1000 and multipying by 100 doesn't give you a percentage).

This adds a `count` metric that reports the total time spent reading and
writting between check runs, instead of a percentage.