add labels from `self.SAMPLE_LABELS` to container status metrics
Users rely on label_join to associate labels from the `kube_pod_labels`
with other `kube_pod_*` metrics. We were previously dropping these
labels for these two metrics. This ensures that labels associated with
pods end up on the `kubernetes_state.container.status_report.*` metrics.

This is similar to how `SAMPLE_LABELS` are processes for other resources
in the check.