datadog_checks_dev: add env management debug logging
When a container is failing to start due to a problem in its startup (i.e. postgres container startup scripts) it's difficult to figure out what's happening because the logs don't appear anywhere. In this case it's easiest to manually run docker-compose but you may not know the exact set of environment variables or other args used by the test framework to start it up. To fix, add the following logging to make it easy to run the docker-compose command manually if necessary:
* add debug logging to `dev/subprocess:run_command` so we can see the full command being run
* add debug logging to `dev/structurs:EnvVars` to log out the full environment being used

If `dev/conditions:WaitFor` is failing it causes the test to be stuck in an extended retry loop. Add debug logging to make it easier to see what the problem is quickly.