add custom check for `cronjob.complete`
This is similar to the current `job.complete` check, but tracks the last
applicable timestamp for the CronJob. If the job with the latest
timestamp is complete, then the check is OK. If the job with the latest
timestamp is failed, then the check is CRITICAL. Otherwise the check is
UNKNOWN.

The existing `job.complete` check is difficult to use for CronJobs as it
will track all existing jobs associated with the CronJob. Due to the
fact that CronJobs often leave around their failed executions, it is
common to have multiple failing/succeeding checks coexisting at the same
time.

Similarly it is difficult to work with the failed/success counts fron
CronJobs as they track all failures, regardless of whether the job
eventually succeeds due to retries. Due to the non standard cadence of
CronJobs, it is often difficult to write suitable monitors using only
failure counts.