postgres exclude EXPLAIN queries from pg_stat_statements
Update the collection of postgres query metrics to exclude `EXPLAIN` queries.

If deep database monitoring is enabled and we are collecting execution plans, then there will be at least one tracked EXPLAIN statement for every unique normalized query we encounter. Since these statements are not used in the product we exclude them from collection.

This updates postgres to have the same behavior as mysql, where we do the same exclusion for query metrics.