log execution plan collection failure at debug level
Execution plans collection is expected to fail in many cases, like if a client is using the extended query protocol meaning we won't get the original parameters from `pg_stat_activity`. To avoid spamming the logs, update this message to be logged only at the debug level.

This now matches what we do for mysql.