Capture Vault route metrics
The Vault route metrics were in the METRIC_MAP and therefore were not
being ingested by Datadog. Because the route metrics exposed by Vault
include the Vault mount path they must be transformed and so the metric
will not be a 1-to-1 rename therefore the metrics are appended to
the METRIC_MAP array.