Add Vault route metrics to be fetched from the Prometheus endpoint (#9612)
* Fix Vault LDAP rollback compat metrics

All Vault mounts have a trailing `/` which is translated to a `_` in the
Prometheus metric name.

This also updates the _sum and _count metrics test data used.

* Capture Vault route metrics

The Vault route metrics were in the METRIC_MAP and therefore were not
being ingested by Datadog. Because the route metrics exposed by Vault
include the Vault mount path they must be transformed and so the metric
will not be a 1-to-1 rename therefore the metrics are appended to
the METRIC_MAP array.