Collect query metrics & plans for Database Monitoring
Adds the [Database Monitoring](https://www.datadoghq.com/product/database-monitoring/) beta for SQL Server, enabling collection of query metrics and execution plans.

**How it works**

If enabled (via `dbm: true`), a python thread is launched during a regular check run:

* collects query metrics from [dm_exec_query_stats](https://docs.microsoft.com/en-us/sql/relational-databases/system-dynamic-management-views/sys-dm-exec-query-stats-transact-sql?view=sql-server-ver15)
* collects an execution plan from [dm_exec_query_plan](https://docs.microsoft.com/en-us/sql/relational-databases/system-dynamic-management-views/sys-dm-exec-query-plan-transact-sql?view=sql-server-ver15) for each query that is active (has changes in query metrics).

Note that this approach differs from how we do it for Postgres and MySQL. Here don't need to collect plans via EXPLAIN because SQL Server makes them available via the Query Plan Cache. The reason plan collection is done following query metric collection is that we need the query metrics to tell us which plans are "active". If a plan is inactive then we don't need to ingest it.