kubelet: don't call the tagger for pods not running
Calling the tagger when an entity does not exist in its store is an
expensive operation, since it'll cause all of its collectors to perform
a remote fetch. Since we know that the tagger does not store pods that
aren't running, calling it for pods that we know not to be running
yet/anymore is waste.

This change stores all pods in PodListUtils, and the summary (that
otherwise has no way of knowing a pod's status) checks to see if the pod
is present and its phase is "Running".