support non-executable files during pipeline setup
If a file in the pipeline setup scripts directory starts with an underscore, then it is treated as a "non-executable" file and it will not be executed.

This is required for https://github.com/DataDog/integrations-core/pull/10637 where a large SQL setup script is needed to be referenced by one of the other scripts, but the SQL script itself must not be executed.