improve sqlserver agent query performance for DBM metrics query
Follow-up to https://github.com/DataDog/integrations-core/pull/10763

Move the filter to the common table expression to avoid temporarily materializing a large result set only to filter out most of it in the following query.

Also add a `TOP {}` safety limit to the query, defaulting to 10k rows to make sure the default behavior is safe and will not cause excessively slow queries in case the database has a large number of plans in the plan cache.