use self healing connection in tests (#10893)
* use self healing connection in tests

Use a new "self-healing connection" (instead of a raw pyodbc connection) for statement metrics & plans tests. There is some inexplicable flakiness that happens during these tests that causes a connection to enter a bad state and no longer be able to make queries reliably. This ensures the connection can be recreated automatically in order to allow the test to proceed.

* raise default timeout

* pin versions