improve internal explain error troubleshooting metrics (#10933)
Update error tags on internal `dd.postgres.statement_samples.error` metrics to all follow the same format. `error:explain-{code}-{message}`, regardless of whether they're a cached or new error. Some errors were failing to report the message when they were cached instead of new.

This ensures we get an accurate representation in the metrics as to which errors are being reported for all the events being ingested.

This also adds a new `dd.postgres.run_explain.error` metric which is only reported when an explain is actually run on the DB and fails. This metric can be compared to the cached explain error metrics to see what fraction of the explain errors are being cached.