run full test suite for every supported Windows SQL Server driver
Remove the limited "can we connect to SQL Server using all supported drivers?" test in favor of running the full Windows test suite once per supported driver: SQLOLEDB, SQLNCLI11, MSOLEDBSQL, odbc.

We've had bugs in the past that are specific to a single driver and this will ensure that all supported drivers are fully tested.