disable modified rowcounts in result sets for all connections
If a stored procedure uses a temporary table then it will, by default, return the number of rows affected as part of the result set. If such a procedure is called (i.e. as part of `custom_queries`) then the query will fail because neither the `pyodbc` nor `adodbapi` libraries are able to deal with this correctly.

For example, adodbapi throws the following error: `fetch() on closed connection or empty query set`

If the procedure sets `SET NOCOUNT ON` internally then the query will succeed. To ensure the agent is able to call all procedures regardless of whether this has been set in the procedure or not, we'll now set `SET NOCOUNT ON` by default on all connections.